+++ src/decllist.h
@@ -23,3 +23,5 @@ void printDeclList(char *id);
 
 OPER *getOper(char *id);
 void addOper(char *id, int preced, ASS_TYPE assoc);
+
+Vector decl_get_ids();